+++ client/src/components/Profile/ChangePassword/ChangePassword.css
@@ -1,19 +1,34 @@
 #changepw-container {
     display: flex;
-    flex-direction: row;
+    flex-direction: column;
     align-items: center;
     justify-content: center;
+
+    margin-top: 1em;
 }
 
 #changepw-form-container {
     min-width: 120px;
     display: flex;
-    flex-direction: row;
+    flex-direction: column;
+}
+
+#changepw-title {
+    margin-bottom: 1em;
 }
 
 
 .changepw-form-field {
     text-align: center;
+    padding-bottom: 4px !important;
+}
+
+.changepw-hidden-field {
+    display: none !important;
+}
+
+#changepw-button {
+    margin-top: 10px;
 }
 
 ::placeholder {
